Started new branch "cproto"
In this branch we tried to use cproto to get function prototypes
from .c and .h files. All the function prototypes are in functions.def
and the memtrace plugin parse that file to extract function names, that
are used to update old function names in call instructions.
File functions.def has still to be generated manually, but we need to
automatically generate it in rootsim-cc.in.

Signed-off-by: Emanuele De Santis <trunk96@gmail.com>